get rid of NVIM_LISTEN_ADDRESS
it has been deprecated for a while, and new versions of nvim won't even
launch if it is set but the file already exists (even if it's a dead
socket)

Instead, I wrote a lil lua script to start the server once vim has
started. It also handles the case where there's a dead socket left over.